requirements.txt
xflow.framework; python_version >= '3.10'
typing_extensions; python_version >= '3.10'